+++ wcfsetup/install/files/lib/page/UsersOnlineListPage.class.php
@@ -35,6 +35,11 @@ class UsersOnlineListPage extends SortablePage {
 	/**
 	 * @inheritDoc
 	 */
+	public $itemsPerPage = 100;
+	
+	/**
+	 * @inheritDoc
+	 */
 	public $defaultSortField = USERS_ONLINE_DEFAULT_SORT_FIELD;
 	
 	/**
@@ -137,8 +142,7 @@ class UsersOnlineListPage extends SortablePage {
 	 * @inheritDoc
 	 */
 	protected function readObjects() {
-		$this->objectList->sqlLimit = 0;
-		if ($this->sqlOrderBy) $this->objectList->sqlOrderBy = ($this->sortField == 'lastActivityTime' ? 'session.' : '').$this->sqlOrderBy;
-		$this->objectList->readObjects();
+		if ($this->sqlOrderBy) $this->sqlOrderBy = ($this->sortField == 'lastActivityTime' ? 'session.' : '').$this->sqlOrderBy;
+		parent::readObjects();
 	}
 }