+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 * `\wcf\system\condition\MultiPageControllerCondition` has been replaced by `wcf\system\condition\page\MultiPageCondition`
 * Added special CSS class to cookie policy notice (`cookiePolicyNotice`)
 * Improved confirmation messages (`<span class="confirmationObject">...</span>`)
+* Added users online list pagination
 
 #### CMS